base/requirements.txt
Django~=4.1.3
djangorestframework~=3.14.0
djoser~=2.0.1
django-filter~=22.1
djangorestframework-simplejwt~=4.8.0
psycopg2-binary~=2.9.5
